FIX: Errors in google oauth2 implementation
LINT: Some warnings
