+++ v2/westack/routing.go
@@ -1,7 +1,6 @@
 package westack
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
 	"log"
@@ -13,12 +12,9 @@ import (
 	"github.com/gofiber/fiber/v2"
 	"github.com/golang-jwt/jwt"
 	"go.mongodb.org/mongo-driver/bson/primitive"
-	"golang.org/x/oauth2"
-	"golang.org/x/oauth2/google"
 
 	wst "github.com/fredyk/westack-go/v2/common"
 	"github.com/fredyk/westack-go/v2/model"
-	"github.com/fredyk/westack-go/v2/utils"
 )
 
 func (app *WeStack) loadNotFoundRoutes() {
@@ -43,7 +39,7 @@ func (app *WeStack) loadModelsFixedRoutes() error {
 
 		e, err := casbin.NewEnforcer(*loadedModel.CasbinModel, *loadedModel.CasbinAdapter, app.debug)
 		if err != nil {
-			return fmt.Errorf("could not create casbin enforcer: %v", err)
+			return fmt.Errorf("could not create casbin enforcer: %w", err)
 		}
 
 		loadedModel.Enforcer = e
@@ -51,19 +47,19 @@ func (app *WeStack) loadModelsFixedRoutes() error {
 		e.EnableAutoSave(true)
 		e.AddFunction("isOwner", casbinOwnerFn(loadedModel))
 
-		err = addDefaultCasbinRoles(app, err, e)
+		err = addDefaultCasbinRoles(app, e)
 		if err != nil {
 			return err
 		}
 
 		err = e.SavePolicy()
 		if err != nil {
-			return fmt.Errorf("could not save policy: %v", err)
+			return fmt.Errorf("could not save policy: %w", err)
 		}
 
 		err = e.LoadPolicy()
 		if err != nil {
-			return fmt.Errorf("could not load policy: %v", err)
+			return fmt.Errorf("could not load policy: %w", err)
 		}
 
 		if app.debug {
@@ -78,7 +74,7 @@ func (app *WeStack) loadModelsFixedRoutes() error {
 			}
 			err = os.WriteFile(fmt.Sprintf("%v/%v.casbin.dump.conf", modelsDumpDir, loadedModel.Name), []byte(text), 0600)
 			if err != nil {
-				return fmt.Errorf("could not write casbin dump: %v", err)
+				return fmt.Errorf("could not write casbin dump: %w", err)
 			}
 		}
 
@@ -322,229 +318,13 @@ func mountAccountModelFixedRoutes(loadedModel *model.StatefulModel, app *WeStack
 		},
 	})
 
-	// oauth2 client
-	googleOauthConfig := &oauth2.Config{
-		ClientID:     app.Viper.GetString("oauth2.google.clientID"),
-		ClientSecret: app.Viper.GetString("oauth2.google.clientSecret"),
-		RedirectURL:  fmt.Sprintf("%s%s/oauth/google/callback", app.Viper.GetString("oauth2.publicOrigin"), loadedModel.BaseUrl),
-		Scopes:       []string{"https://www.googleapis.com/auth/userinfo.email"},
-		Endpoint:     google.Endpoint,
-	}
+	mountOauthRoutes(app, loadedModel, systemContext)
 
-	// google login endpoint
-	loadedModel.RemoteMethod(func(eventContext *model.EventContext) error {
+}
 
-		oauthStateString := wst.CreateOauthStateString()
-		cookie := ""
-		if eventContext.Ctx.Cookies("SSID") != "" {
-			cookie = eventContext.Ctx.Cookies("SSID")
-		} else {
-			cookie = wst.GenerateCookie()
-			eventContext.Ctx.Cookie(&fiber.Cookie{
-				Name:  "SSID",
-				Value: cookie,
-			})
-		}
-		utils.RegisterOauthState(cookie, oauthStateString)
-
-		url := googleOauthConfig.AuthCodeURL(oauthStateString, oauth2.AccessTypeOffline)
-		return eventContext.Ctx.Redirect(url)
-	}, model.RemoteMethodOptions{
-		Name:        string(wst.OperationNameGoogleLogin),
-		Description: "Logins with Google",
-		Http: model.RemoteMethodOptionsHttp{
-			Path: "/oauth/google",
-			Verb: "get",
-		},
-	})
-
-	// google callback endpoint
-	loadedModel.RemoteMethod(func(eventContext *model.EventContext) error {
-
-		cookie := eventContext.Ctx.Cookies("SSID")
-		oauthState, ok := utils.GetOauthState(cookie)
-		if !ok {
-			return wst.CreateError(fiber.ErrForbidden, "ERR_MISSING_OAUTH_STATE", fiber.Map{"message": "missing oauth state"}, "Error")
-		}
-
-		receivedState := eventContext.Query.GetString("state")
-		if oauthState != receivedState {
-			return wst.CreateError(fiber.ErrForbidden, "ERR_INVALID_OAUTH_STATE", fiber.Map{"message": "invalid oauth state"}, "Error")
-		}
-
-		oauthCode := eventContext.Query.GetString("code")
-		token, err := googleOauthConfig.Exchange(eventContext.Ctx.Context(), oauthCode)
-		if err != nil {
-			return wst.CreateError(fiber.ErrForbidden, "ERR_OAUTH_EXCHANGE", fiber.Map{"message": "oauth exchange failed: " + err.Error()}, "Error")
-		}
-
-		userInfo, err := googleOauthConfig.Client(eventContext.Ctx.Context(), token).Get("https://www.googleapis.com/oauth2/v3/userinfo")
-		if err != nil {
-			return wst.CreateError(fiber.ErrForbidden, "ERR_OAUTH_USERINFO", fiber.Map{"message": "failed to get user info: " + err.Error()}, "Error")
-		}
-
-		defer userInfo.Body.Close()
-
-		type userInfoResponse struct {
-			Email string `json:"email"`
-		}
-
-		var userInfoData userInfoResponse
-		err = json.NewDecoder(userInfo.Body).Decode(&userInfoData)
-		if err != nil {
-			return wst.CreateError(fiber.ErrForbidden, "ERR_OAUTH_USERINFO_DECODE", fiber.Map{"message": "failed to decode user info: " + err.Error()}, "Error")
-		}
-
-		// check if userCredentials exists
-		userCredentials, err := app.accountCredentialsModel.FindOne(&wst.Filter{
-			Where: &wst.Where{
-				"email":    userInfoData.Email,
-				"provider": ProviderGoogleOAuth2,
-			},
-			Include: &wst.Include{
-				{
-					Relation: "account",
-				},
-			},
-		}, systemContext)
-
-		if err != nil {
-			fmt.Printf("[DEBUG] Error while fetching credentials by email-provider: %v\n", err)
-			return err
-		}
-
-		var account *model.StatefulInstance
-		var accountId string
-
-		if userCredentials == nil {
-			// search by password
-			userCredentials, err = app.accountCredentialsModel.FindOne(&wst.Filter{
-				Where: &wst.Where{
-					"email": userInfoData.Email,
-					"$or": []wst.M{
-						{"provider": ProviderPassword},
-						{"password": wst.M{"$exists": true}},
-					},
-				},
-				Include: &wst.Include{
-					{
-						Relation: "account",
-					},
-				},
-			}, systemContext)
-
-			if err != nil {
-				fmt.Printf("[DEBUG] Error while fetching credentials by email-password: %v\n", err)
-				return err
-			}
-
-			if userCredentials == nil {
-
-				// create new account
-				fmt.Printf("[DEBUG] Creating new account for email: %v\n", userInfoData.Email)
-				createdAccount, err := loadedModel.Create(wst.M{
-					"email":         userInfoData.Email,
-					"emailVerified": true,
-					"provider":      ProviderGoogleOAuth2,
-				}, systemContext)
-
-				if err != nil {
-					fmt.Printf("[DEBUG] Error while creating account: %v\n", err)
-					return err
-				}
-
-				account = createdAccount.(*model.StatefulInstance)
-				accountId = account.GetString("id")
-
-			} else {
-
-				accountId = userCredentials.GetString("accountId")
-
-				account = userCredentials.GetOne("account").(*model.StatefulInstance)
-
-			}
-
-			// create new credentials
-			fmt.Printf("[DEBUG] Creating new credentials for email: %v\n", userInfoData.Email)
-			_, err = app.accountCredentialsModel.Create(wst.M{
-				"accountId":    accountId,
-				"email":        userInfoData.Email,
-				"provider":     ProviderGoogleOAuth2,
-				"accessToken":  token.AccessToken,
-				"refreshToken": token.RefreshToken,
-				"expiry":       token.Expiry,
-				"tokenType":    token.TokenType,
-				"scope":        token.Extra("scope"),
-			}, systemContext)
-
-			if err != nil {
-				fmt.Printf("[DEBUG] Error while creating credentials: %v\n", err)
-				return err
-			}
-
-		} else {
-			account = userCredentials.GetOne("account").(*model.StatefulInstance)
-
-			// update credentials
-			fmt.Printf("[DEBUG] Updating credentials for email: %v\n", userInfoData.Email)
-			_, err = userCredentials.UpdateAttributes(wst.M{
-				"accessToken": token.AccessToken,
-				"expiry":      token.Expiry,
-				"tokenType":   token.TokenType,
-				"scope":       token.Extra("scope"),
-			}, systemContext)
-
-			if err != nil {
-				fmt.Printf("[DEBUG] Error while updating credentials: %v\n", err)
-				return err
-			}
-
-		}
-
-		roleNames := []string{"USER"}
-
-		roleContext := &model.EventContext{
-			BaseContext:            systemContext,
-			DisableTypeConversions: true,
-		}
-
-		roleEntries, _ := app.roleMappingModel.FindMany(&wst.Filter{Where: &wst.Where{
-			"principalType": "USER",
-			"$or": []wst.M{
-				{
-					"principalId": accountId,
-				},
-				{
-					"principalId": account.Id,
-				},
-			},
-		}, Include: &wst.Include{{Relation: "role"}}}, roleContext).All()
-
-		for _, roleEntry := range roleEntries {
-			role := roleEntry.GetOne("role")
-			roleNames = append(roleNames, role.ToJSON()["name"].(string))
-		}
-
-		ttl := 30 * 86400.0
-		bearer := model.CreateBearer(eventContext.ModelID, float64(time.Now().Unix()), ttl, roleNames)
-		// sign the bearer
-		jwtToken := jwt.NewWithClaims(jwt.SigningMethodHS256, bearer.Claims)
-		tokenString, err := jwtToken.SignedString(loadedModel.App.JwtSecretKey)
-		if err != nil {
-			return err
-		}
-
-		return eventContext.Ctx.Redirect(app.Viper.GetString("oauth2.successRedirect") + "?access_token=" + tokenString)
-
-	}, model.RemoteMethodOptions{
-		Name:        string(wst.OperationNameGoogleLoginCallback),
-		Description: "Google OAuth2 callback",
-		Http: model.RemoteMethodOptionsHttp{
-			Path: "/oauth/google/callback",
-			Verb: "get",
-		},
-	})
-
+func verboseRedirect(eventContext *model.EventContext, failureUrl string, err error) error {
+	fmt.Printf("[DEBUG] Redirecting to %v: %v\n", failureUrl, err)
+	return eventContext.Ctx.Redirect(fmt.Sprintf("%v?error=%v", failureUrl, err.Error()))
 }
 
 func mountBaseModelFixedRoutes(app *WeStack, loadedModel *model.StatefulModel) {
@@ -678,7 +458,7 @@ func mountAppDynamicRoutes(loadedModel *model.StatefulModel, app *WeStack) {
 	})
 }
 
-func addDefaultCasbinRoles(app *WeStack, err error, e *casbin.Enforcer) error {
+func addDefaultCasbinRoles(app *WeStack, e *casbin.Enforcer) (err error) {
 	_, err = e.AddRoleForUser("findMany", replaceVarNames("read"))
 	if app.debug {
 		app.logger.Printf("[DEBUG] Added role findMany for user %v, err: %v\n", replaceVarNames("read"), err)
@@ -827,7 +607,7 @@ func (app *WeStack) loadModelsDynamicRoutes() {
 }
 
 func handleEvent(eventContext *model.EventContext, loadedModel *model.StatefulModel, event string) (err error) {
-	if loadedModel.DisabledHandlers[event] != true {
+	if !loadedModel.DisabledHandlers[event] {
 		err = loadedModel.GetHandler(event)(eventContext)
 		if err != nil {
 			return
@@ -866,7 +646,7 @@ func casbinOwnerFn(loadedModel *model.StatefulModel) func(arguments ...interface
 		policyObj := arguments[2]
 
 		if loadedModel.App.Debug {
-			log.Println(fmt.Sprintf("isOwner() of %v ?", policyObj))
+			fmt.Printf("isOwner() of %v ?", policyObj)
 		}
 
 		objId = strings.TrimSpace(objId)
